Cell object for FlexTable

Create a representation of a cell that can be inserted in a FlexRow. For internal usage.

FlexCell(value, colspan = 1, par.properties = parProperties(),
  cell.properties = cellProperties())

Arguments

value

a content value - a value of type character or pot or set_of_paragraphs.

colspan

defines the number of columns the cell should span

par.properties

parProperties to apply to content

cell.properties

cellProperties to apply to content

See also

addFlexTable, addHeaderRow, addFooterRow

Examples

#
FlexCell( value = "Hello" )

#> Hello

FlexCell( value = "Hello", colspan = 3)

#> Hello

FlexCell( "Column 1",
        cell.properties = cellProperties(background.color="#527578")  )

#> Column 1

# define a complex formatted text
mytext = pot("Hello", format = textProperties(color = "blue")
  ) + " " + pot( "world", format = textProperties(font.size = 9)
)
Fcell = FlexCell( mytext, colspan = 4 )

# define two paragraph and put them in a FlexCell
mytext1 = pot("Hello", format = textProperties(color = "blue") )
mytext2 = pot( "world", format = textProperties(font.size = 9) )
Fcell = FlexCell( set_of_paragraphs( mytext1, mytext2 ) )