ReporteRs: a package to create document from R

ReporteRs lets you create Microsoft Word, Microsoft PowerPoint and html documents.

Details

To get an r document object:

docx Create a Microsoft Word document object

pptx Create a Microsoft PowerPoint document object

bsdoc Create an HTML document object (deprecated)

The following functions can be used whatever the output format is (docx, pptx, bsdoc).

addTitle Add a title

addFlexTable Add a table (new)

addPlot Add plots

addImage Add external images

addParagraph Add paragraphs of text

addRScript Add an r script

writeDoc Write the document into a file or a directory

ReporteRs comes with an object of class pot to let you handle text output and format. You can associate a text with formats (font size, font color, etc.), with an hyperlink or with a Footnote as a reference note.

ReporteRs comes also with an object of class FlexTable that let you design and format tabular outputs.

Default values:

With ReporteRs, some options can be used to reduce usage of some parameters:

"ReporteRs-default-font" Default font family to use (default to "Helvetica"). This will be used as default values for argument fontname of addPlot and argument font.family of pot.

"ReporteRs-fontsize" Default font size to use (default to 11). This will be used as default values for argument pointsize of addPlot and argument font.size of pot.

"ReporteRs-list-definition" see list.settings.

"ReporteRs-locale.language" language encoding (for html objects). Default to "en".

"ReporteRs-locale.region" region encoding (for html objects). Default to "US".

Note

Examples are in a dontrun section as they are using font that may be not available on the host machine. Default font is Helvetica, it can be modified with option ReporteRs-default-font. To run an example with 'Arial' default font, run first

options("ReporteRs-default-font" = "Arial")

Examples

options("ReporteRs-fontsize"=10, "ReporteRs-default-font"="Helvetica")
numbering.pattern = c( "%1.", "%1. %2.", "%1. %2. %3.",
  "%4.", "%5.", "%6.", "%7.", "%8.", "%9." )

ordered.formats = rep( c( "decimal", "upperRoman", "upperLetter"), 3 )

unordered.formats = rep( c( "square", "disc", "circle"), 3 )

left.indent = seq( from = 0, by = 0.5, length.out = 9)

options("ReporteRs-list-definition" = list(
  ol.left = left.indent,
  ol.hanging = rep( 0.4, 9 ),
  ol.format = ordered.formats,
  ol.pattern = numbering.pattern,
  ul.left = left.indent,
  ul.hanging = rep( 0.4, 9 ),
  ul.format = unordered.formats
  )
)