Create an object representation of a bootstrap html

Create a bsdoc object

bsdoc(title = "untitled",
  list.definition = getOption("ReporteRs-list-definition"), keywords = "",
  description = "", mathjax = FALSE)

Arguments

title

"character" value: title of the document.

list.definition

a list definition to specify how ordered and unordered lists have to be formated. See list.settings. Default to getOption("ReporteRs-list-definition").

keywords

"character" value: keywords metadata value to set in the html page

description

"character" value: description metadata value to set in the html page

mathjax

logical value: if TRUE activate mathjax

Value

an object of class bsdoc.

Details

Several methods can used to send R output into an object of class bsdoc.

addTitle.bsdoc add titles

addParagraph.bsdoc add text

addPlot.bsdoc add plots

addFlexTable.bsdoc add tables. See FlexTable

addImage.bsdoc add external images

addRScript.bsdoc add highlighted r script

Once object has content, user can write the docx into a ".html" file, see writeDoc.bsdoc.

Note

bsdoc objects are deprecated and will be deleted in the next version.