Cell formatting properties

Create a cellProperties object that describes cell formatting properties. This objects are used by FlexTable.

cellProperties(padding, border.width, border.style, border.color, border.bottom,
  border.left, border.top, border.right, border.bottom.color = "black",
  border.bottom.style = "solid", border.bottom.width = 1,
  border.left.color = "black", border.left.style = "solid",
  border.left.width = 1, border.top.color = "black",
  border.top.style = "solid", border.top.width = 1,
  border.right.color = "black", border.right.style = "solid",
  border.right.width = 1, vertical.align = "middle", padding.bottom = 0,
  padding.top = 0, padding.left = 0, padding.right = 0,
  background.color = "transparent", text.direction = "lrtb")

# S3 method for cellProperties
chprop(object, border.bottom, border.left, border.top,
  border.right, padding, border.bottom.color, border.bottom.style,
  border.bottom.width, border.left.color, border.left.style, border.left.width,
  border.top.color, border.top.style, border.top.width, border.right.color,
  border.right.style, border.right.width, vertical.align, padding.bottom,
  padding.top, padding.left, padding.right, background.color, text.direction,
  ...)

# S3 method for cellProperties
print(x, ...)

Arguments

padding

cell padding - 0 or positive integer value. Argument padding overwrites arguments padding.bottom, padding.top, padding.left, padding.right.

border.width

border width - 0 or positive integer value. Argument border.width overwrites arguments border.bottom.width, border.top.width , border.left.width, border.right.width.

border.style

border style - a single character value, expected value is one of "none" , "solid", "dotted", "dashed". Argument border.style overwrites arguments border.bottom.style, border.top.style, border.left.style , border.right.style.

border.color

border color - a single character value specifying a valid color (e.g. "#000000" or "black"). Argument border.color overwrites arguments border.bottom.color , border.top.color, border.left.color, border.right.color.

border.bottom

borderProperties for bottom border. overwrite all border.bottom.* if specified.

border.left

borderProperties for left border. overwrite all border.left.* if specified.

border.top

borderProperties for top border. overwrite all border.top.* if specified.

border.right

borderProperties for right border. overwrite all border.right.* if specified.

border.bottom.color

border bottom color - a single character value specifying a valid color (e.g. "#000000" or "black").

border.bottom.style

border bottom style - a single character value, expected value is one of "none", "solid", "dotted", "dashed".

border.bottom.width

border bottom width - 0 or positive integer value

border.left.color

border left color - a single character value specifying a valid color (e.g. "#000000" or "black").

border.left.style

border left style - a single character value, expected value is one of "none", "solid", "dotted", "dashed".

border.left.width

border left width - 0 or positive integer value

border.top.color

border top color - a single character value specifying a valid color (e.g. "#000000" or "black").

border.top.style

border top style - a single character value, expected value is one of "none", "solid", "dotted", "dashed".

border.top.width

border top width - 0 or positive integer value

border.right.color

border right color - a single character value specifying a valid color (e.g. "#000000" or "black").

border.right.style

border right style - a single character value, expected value is one of "none", "solid", "dotted", "dashed".

border.right.width

border right width - 0 or positive integer value

vertical.align

cell content vertical alignment - a single character value , expected value is one of "center" or "top" or "bottom"

padding.bottom

cell bottom padding - 0 or positive integer value.

padding.top

cell top padding - 0 or positive integer value.

padding.left

cell left padding - 0 or positive integer value.

padding.right

cell right padding - 0 or positive integer value.

background.color

cell background color - a single character value specifying a valid color (e.g. "#000000" or "black").

text.direction

cell text rotation - a single character value, expected value is one of "lrtb", "tbrl", "btlr".

object

cellProperties object to update

...

further arguments - not used

x

cellProperties object to print

Details

Default values are:

border.bottom.color "black"

border.bottom.style "solid"

border.bottom.width 1

border.left.color "black"

border.left.style "solid"

border.left.width 1

border.top.color "black"

border.top.style "solid"

border.top.width 1

border.right.color "black"

border.right.style "solid"

border.right.width 1

vertical.align "middle"

padding.bottom 1

padding.top 1

padding.left 1

padding.right 1

background.color "white"

text.direction "lrtb"

Get a modified version of a cellProperties with chprop.

See also

borderProperties, FlexTable, shortcut_properties

Examples

# cellProperties examples -------
cellProp01 = cellProperties( border.color = "gray", border.width = 2 )
cellProp02 = cellProperties(border.left.width = 0, border.right.width = 0
  , border.bottom.width = 2, border.top.width = 0
  , padding.bottom = 2, padding.top = 2
  , padding.left = 2, padding.right = 2 )

# chprop usage example ------
cellProp = cellProperties()

cellProp01 = chprop( cellProp, border.bottom.color = "#8A949B" )
cellProp02 = chprop( cellProp, border.right.color = "#8A949B" )
cellProp03 = chprop( cellProp, border.left.color = "#8A949B" )
cellProp04 = chprop( cellProp, border.top.color = "#8A949B" )
cellProp05 = chprop( cellProp, border.color = "#8A949B" )

cellProp06 = chprop( cellProp, border.bottom.width = 2 )
cellProp07 = chprop( cellProp, border.left.width = 2 )
cellProp08 = chprop( cellProp, border.top.width = 2 )
cellProp09 = chprop( cellProp, border.right.width = 2 )
cellProp10 = chprop( cellProp, border.width = 2 )

cellProp11 = chprop( cellProp, padding.bottom = 5 )
cellProp12 = chprop( cellProp, padding.top = 5 )
cellProp13 = chprop( cellProp, padding.left = 5 )
cellProp14 = chprop( cellProp, padding.right = 5 )
cellProp15 = chprop( cellProp, padding = 5 )

cellProp16 = chprop( cellProp,
        border.bottom = borderProperties( style = "dotted" ) )
cellProp17 = chprop( cellProp, border.left.style = "dotted" )
cellProp18 = chprop( cellProp, border.top.style = "dotted" )
cellProp19 = chprop( cellProp, border.right.style = "dotted" )
cellProp20 = chprop( cellProp, border.style = "dotted" )

cellProp21 = chprop( cellProp, vertical.align = "middle" )
cellProp22 = chprop( cellProp, background.color = "#517281" )

cellProp23 = chprop( cellProp, background.color = "#517281"
  , border.color = "#F37257", border.width = 2  )