map titles styles

Set manually headers'styles of a docx object

map_title(doc, stylenames)

Arguments

doc

docx object to be used with map_title.

stylenames

existing styles (character vector) where first element represents the style to use for title 1, second element represents the style to use for title 2, etc.

Details

Function addTitle need to know which styles are corresponding to which title level (1 ; 1.1 ; 1.1.1 ; etc.). When template is read, function docx try to guess what are theses styles. If he do not succeed, an error occured saying 'You must defined header styles via map_title first.'. In that case, run styles(...) to see what are available styles, then map_title to indicate which available styles are meant to be used as header styles.

See also

docx,styles.docx,addTitle.docx

Examples

doc.filename = "addImage_example.docx"
doc <- docx()
doc = map_title(doc, stylenames = c("Titre1", "Titre2", "Titre3",
  "Titre4", "Titre5", "Titre6", "Titre7", "Titre8", "Titre9" ) )