Piece of Text (formated text)

Create an object with a text to display and its formatting properties.

pot(value = "", format = textProperties(), hyperlink, footnote)

# S3 method for pot
print(x, ...)

# S3 method for pot
as.character(x, ...)

# S3 method for pot
knit_print(x, ...)

Arguments

value

text value or a value that has a format method returning character value.

format

formatting properties (an object of class textProperties).

hyperlink

a valid url to use as hyperlink when clicking on value.

footnote

a Footnote object.

x

a pot object

...

further arguments, not used.

Details

a pot (piece of text) is a convenient way to define a paragraph of text where some text are not all formated the same.

A pot can be associated with an hyperlink.

A pot can be associated with a Footnote. Note that footnotes can not be inserted in a pptx object.

See also

addParagraph, Footnote , +.pot, pot_img,

Examples

# "My tailor is rich" with formatting on some words
pot1 <- pot("My tailor", textProperties(color = "red" ) ) +
  " is " + pot("rich", textProperties(shading.color = "red",
     font.weight = "bold" ) )

# "Cats and dogs" with formatting on some words
pot2 = pot("Cats", textProperties(color = "red" ) ) +
  " and " +
  pot("dogs", textProperties( color = "blue" ),
      hyperlink = "http://www.wikipedia.org/" )